writing-practice/requirements.txt
gradio>=4.12.0
boto3>=1.34.0
python-dotenv>=1.0.0
manga-ocr>=0.1.11
Pillow>=10.0.0
requests>=2.31.0
python-json-logger>=2.0.7
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
